+++ src/app/(components)/admin/userForms/AddUser.jsx
@@ -5,6 +5,7 @@ import { COLOURS } from "@/app/constants";
 import Link from "next/link";
 import Button from "../../buttons/Button";
 import { v4 as uuidv4 } from "uuid";
+import userRoleInfo from "./userRoleInfo.json";
 
 const AddUser = () => {
   const generatePassword = () => {
@@ -208,19 +209,31 @@ const AddUser = () => {
           <div style={{ marginBottom: "20px" }}>
             {formData.role === "admin" && (
               <div style={userInformationStyle}>
-                Admin-specific content goes here
+                <ul>
+                  {userRoleInfo.admin.map((item, index) => (
+                    <li key={index}>{item}</li>
+                  ))}
+                </ul>
               </div>
             )}
 
             {formData.role === "manager" && (
               <div style={userInformationStyle}>
-                Manager-specific content goes here
+                <ul>
+                  {userRoleInfo.manager.map((item, index) => (
+                    <li key={index}>{item}</li>
+                  ))}
+                </ul>
               </div>
             )}
 
             {formData.role === "user" && (
               <div style={userInformationStyle}>
-                User-specific content goes here
+                <ul>
+                  {userRoleInfo.user.map((item, index) => (
+                    <li key={index}>{item}</li>
+                  ))}
+                </ul>
               </div>
             )}
           </div>