+++ src/app/(components)/admin/userForms/ConfirmDeleteUserModal.jsx
@@ -25,6 +25,7 @@ function ConfirmDeleteUserModal({ user, onConfirm, onCancel }) {
 
         <div
           style={{
+            paddingTop: "20px",
             paddingLeft: "20px",
             paddingRight: "20px",
             paddingBottom: "20px",