+++ src/app/(components)/forms/modals/PasswordUpdatedModal.jsx
@@ -25,6 +25,7 @@ function PasswordUpdatedModal({ onClose }) {
 
         <div
           style={{
+            paddingTop: "20px",
             paddingLeft: "20px",
             paddingRight: "20px",
             paddingBottom: "20px",