+++ src/app/(components)/projects/manager/project-list.jsx
@@ -235,7 +235,10 @@ function ManagersProjectsList() {
         </div>
       </div>
 
-      <div className="project-list-table-rows">
+      <div
+        className="project-list-table-rows"
+        style={{ marginBottom: "120px" }}
+      >
         <ul>
           {filteredProjects.map((project) => (
             <li