+++ src/app/(components)/timesheet/modal/ConfirmDeleteModal.jsx
@@ -25,6 +25,7 @@ function ConfirmDeleteModal({ timesheet, onConfirm, onCancel }) {
 
         <div
           style={{
+            paddingTop: "20px",
             paddingLeft: "20px",
             paddingRight: "20px",
             paddingBottom: "20px",