+++ changelog.md
@@ -1,5 +1,18 @@
 ## [0.21] - 26/03/2024
 
+BT-57 Clean up functions
+
+## Added
+
+- Ability to delete projects from the system
+
+## Fixes
+
+- Modal padding in BT-62
+- User permissions shown in AddUser and EditUser
+
+## [0.21] - 26/03/2024
+
 -- no ticket
 
 ## Fixes